Moved mouse tests to parallel
